+++ implicit_solver/lib/objects/components/constraints/bending.py
@@ -34,7 +34,7 @@ class Bending(ConstraintBase):
         x0, v0 = na.node_xv(scene.dynamics, node_ids[0])
         x1, v1 = na.node_xv(scene.dynamics, node_ids[1])
         x2, v2 = na.node_xv(scene.dynamics, node_ids[2])
-        self.rest_angle = math2D.angle(x0, x1, x2)
+        self.rest_angle = np.float64(math2D.angle(x0, x1, x2))
         self.node_IDs = np.copy(node_ids)
 
     @classmethod
